# docs/endpoints/openapi-systems-ops.yaml
openapi: 3.0.3
info:
  title: Atlas API — Systems Ops
  version: '1.0.0'
  description: |
    Contrato OpenAPI ULTRA DETALLADO, generado leyendo directamente el código fuente
    (controladores, schemas Zod, mappers, servicios y repositorios), del módulo
    `systems-ops` del backend Atlas: catálogo técnico de endpoints/herramientas/entidades
    de datos, log de acciones (auditoría HTTP), cola de revisión de gobierno de datos,
    perfiles y runs de stress testing, y suites/pasos/runs de pruebas E2E internas.

    ## Envolturas globales (aplican a TODOS los endpoints de este documento)

    Este backend registra dos providers globales en `src/app.module.ts`
    (`APP_INTERCEPTOR` y `APP_FILTER`) que envuelven **toda** respuesta, éxito o error.
    Los schemas de `data` documentados endpoint por endpoint abajo son el contenido
    de la propiedad `data`, **no** el body HTTP completo.

    **Éxito** (`ResponseInterceptor`, `src/common/interceptors/response.interceptor.ts`):
    ```json
    { "requestId": "corr-id-o-undefined", "data": { "...": "..." }, "timestamp": "2026-07-07T12:00:00.000Z" }
    ```

    **Error** (`HttpExceptionFilter`, `src/common/filters/http-exception.filter.ts`):
    ```json
    { "requestId": "corr-id-o-undefined", "error": { "code": "VALIDATION_ERROR", "message": "..." }, "timestamp": "2026-07-07T12:00:00.000Z" }
    ```
    `error.code` se deriva del status HTTP: `400→VALIDATION_ERROR`, `401→UNAUTHORIZED`,
...
    de query/params/body.

...

...
    ```ts
...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /systems/review-queue:
    get:
      operationId: getReviewQueue
      tags: [SystemsReview]
      summary: Cola unificada de elementos del catálogo pendientes de revisión
      description: 'Envuelve cada colección como `{items, total}` (SIN `page`/`limit`/`totalPages`) — distinto del envelope `{items, meta}` usado en el resto del módulo.'
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [all, endpoints, data_entities, data_impacts, field_impacts, data_column_impacts, tool_requirements]
            default: all
          description: Qué colección(es) incluir en la respuesta.
        - $ref: '#/components/parameters/ModuleQuery'
        - name: reviewStatus
          in: query
          schema:
            type: string
            enum: [AUTO_DETECTED, NEEDS_REVIEW, APPROVED, REJECTED]
            default: NEEDS_REVIEW
          description: Estado de revisión a filtrar en todas las colecciones.
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/LimitQuery'
      responses:
        '200':
          description: Cola de revisión agregada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_ReviewQueue'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /systems/endpoints/{endpointId}/review:
    patch:
      operationId: reviewEndpoint
      tags: [SystemsReview]
      summary: Aprobar, rechazar o marcar para revisión un endpoint del catálogo
      description: Usa `@CurrentUser()` para registrar `updatedBy`.
      parameters:
        - $ref: '#/components/parameters/EndpointIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReviewDecisionRequest' }
      responses:
        '200':
          description: Endpoint actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_EndpointDetail_EndpointOnly'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: SYSTEM_ENDPOINT_NOT_FOUND
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /systems/tools/requirements/{requirementId}/review:
    patch:
      operationId: reviewToolRequirement
      tags: [SystemsReview]
      summary: Revisar un requisito de herramienta inferido/manual para un endpoint
      parameters:
        - $ref: '#/components/parameters/RequirementIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReviewDecisionRequest' }
      responses:
        '200':
          description: Requisito de herramienta actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_ToolRequirement'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: SYSTEM_TOOL_REQUIREMENT_NOT_FOUND
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /systems/data-entities/{entityId}/review:
    patch:
      operationId: reviewDataEntity
      tags: [SystemsReview]
      summary: Revisar (aprobar/rechazar) una entidad de datos del catálogo
      parameters:
        - $ref: '#/components/parameters/EntityIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReviewDecisionRequest' }
      responses:
        '200':
          description: Entidad actualizada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_DataEntity'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: SYSTEM_DATA_ENTITY_NOT_FOUND
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /systems/data-entities/columns/{columnId}/review:
    patch:
      operationId: reviewDataColumn
      tags: [SystemsReview]
      summary: Revisar una columna del catalogo de datos
      description: Reutiliza `ReviewDecisionRequest`; si viene `notes`, se guarda como nota operacional y marca la columna como editada manualmente.
      parameters:
        - name: columnId
          in: path
          required: true
          schema: { type: string, pattern: '^[1-9][0-9]*$' }
          description: Id numerico de la columna en system_data_field_catalog.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReviewDecisionRequest' }
      responses:
        '200':
          description: Columna actualizada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_DataEntityColumn'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: SYSTEM_DATA_COLUMN_NOT_FOUND
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /systems/impact/data/{impactId}/review:
    patch:
      operationId: reviewDataImpact
      tags: [SystemsReview]
      summary: Revisar un impacto de endpoint sobre entidad de datos
      description: A diferencia de `reviewEndpoint`/`reviewDataEntity`, el campo `notes` de `ReviewDecisionRequest` SÍ se persiste aquí si viene definido.
      parameters:
        - name: impactId
          in: path
          required: true
          schema: { type: string, pattern: '^[1-9][0-9]*$' }
          description: Id numérico del impacto de datos.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReviewDecisionRequest' }
      responses:
        '200':
          description: Impacto de datos actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_DataImpact'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: SYSTEM_DATA_IMPACT_NOT_FOUND
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /systems/impact/fields/{fieldImpactId}/review:
    patch:
      operationId: reviewFieldImpact
      tags: [SystemsReview]
      summary: Revisar un impacto de endpoint sobre un campo específico
      description: El campo `notes` de `ReviewDecisionRequest` SÍ se persiste aquí si viene definido.
      parameters:
        - name: fieldImpactId
          in: path
          required: true
          schema: { type: string, pattern: '^[1-9][0-9]*$' }
          description: Id numérico del impacto de campo.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReviewDecisionRequest' }
      responses:
        '200':
          description: Impacto de campo actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_FieldImpact'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: SYSTEM_FIELD_IMPACT_NOT_FOUND
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }

  # ============================================================
  # SystemsStressController
  # ============================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT emitido por `POST /api/v1/internal/auth/login` (o `/api/v1/auth/login` para
        roles no internos). Debe pertenecer a un usuario con rol en:
        system_admin, platform_admin, admin, qa_engineer, devops, risk_analyst,
        compliance_analyst, readonly_auditor.
  responses:
    BadRequest:
      description: Entrada inválida (Zod) o regla de negocio violada — ver detalle por endpoint.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unauthorized:
      description: Falta `Authorization`, formato inválido, token expirado/inválido o revocado.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Forbidden:
      description: 'Rol autenticado no está en la lista permitida: "El usuario autenticado no tiene permiso para esta operación."'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  parameters:
    PageQuery:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
      description: Número de página (entero positivo, coercionado desde string).
    LimitQuery:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      description: Tamaño de página (máx 100).
    ModuleQuery:
      name: module
      in: query
      schema: { type: string, minLength: 1, maxLength: 120 }
      description: Filtra por módulo.
...
    EndpointIdPath:
      name: endpointId
      in: path
      required: true
      schema: { type: string, pattern: '^[1-9][0-9]*$' }
      description: Id numérico del endpoint (representado como texto).
...
    EntityIdPath:
      name: entityId
      in: path
      required: true
      schema: { type: string, pattern: '^[1-9][0-9]*$' }
      description: Id numérico de la entidad de datos.
...
    RequirementIdPath:
      name: requirementId
      in: path
      required: true
      schema: { type: string, pattern: '^[1-9][0-9]*$' }
      description: Id numérico del requisito de herramienta.
...
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        requestId: { type: string, nullable: true, example: 'req-8f3c2a' }
        error:
          type: object
          properties:
            code:
              type: string
              enum: [VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, CONFLICT, INTERNAL_ERROR]
            message: { type: string, example: 'SYSTEM_ENDPOINT_NOT_FOUND' }
          required: [code, message]
        timestamp: { type: string, format: date-time }
      required: [error, timestamp]
...
    Endpoint:
      type: object
      description: Forma producida por `mapEndpoint` (src/modules/systems-ops/systems-ops.mapper.ts).
      properties:
        endpointId: { type: string }
        code: { type: string }
        module: { type: string }
        controllerName: { type: string, nullable: true }
        handlerName: { type: string, nullable: true }
        method: { type: string }
        routePath: { type: string, description: 'Ruta relativa, sin el prefijo /api/v1.' }
        fullPath: { type: string, description: 'Ruta completa con prefijo /api/v1.' }
        routeName: { type: string }
        businessPurpose: { type: string }
        businessAction: { type: string, nullable: true }
        expectedResponseSummary: { type: string, nullable: true }
        expectedStatusCodes: { type: array, items: { type: integer } }
        minPayloadSchema: { type: object }
        queryParamsSchema: { type: object }
        pathParamsSchema: { type: object }
        headersSchema: { type: object }
        requiresAuth: { type: boolean }
        allowedRoles: { type: array, items: { type: string } }
        containsPii: { type: boolean }
        piiFields: { type: array, items: { type: string } }
        riskLevel: { type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL] }
        isDestructive: { type: boolean }
        isReadonly: { type: boolean }
        idempotencyRequired: { type: boolean }
        requiresStressTest: { type: boolean }
        requiresIntegrationTest: { type: boolean }
        isTestableFromPortal: { type: boolean }
        testEnvironmentOnly: { type: boolean }
        ownerTeam: { type: string, nullable: true }
        status: { type: string, example: ACTIVE }
        version: { type: string, example: v1 }
        detectedFrom: { type: string, example: controller }
        confidenceLevel: { type: string, example: MEDIUM }
        reviewStatus: { type: string, enum: [AUTO_DETECTED, NEEDS_REVIEW, APPROVED, REJECTED] }
        sourceFile: { type: string, nullable: true }
        createdAt: { type: string, format: date-time, nullable: true }
        updatedAt: { type: string, format: date-time, nullable: true }
...
    DataEntity:
      type: object
      description: Forma producida por `mapDataEntity`.
      properties:
        entityId: { type: string }
        schemaName: { type: string }
        tableName: { type: string }
        modelName: { type: string, nullable: true }
        entityName: { type: string }
        module: { type: string }
        businessPurpose: { type: string }
        dataOwner: { type: string }
        containsPii: { type: boolean }
        containsFinancialData: { type: boolean }
        containsRiskData: { type: boolean }
        containsLegalData: { type: boolean }
        containsDeviceData: { type: boolean }
        containsLocationData: { type: boolean }
        isAuditCritical: { type: boolean }
        retentionPolicyCode: { type: string, nullable: true }
        status: { type: string }
        detectedFrom: { type: string }
        confidenceLevel: { type: string }
        reviewStatus: { type: string, enum: [AUTO_DETECTED, NEEDS_REVIEW, APPROVED, REJECTED] }
        columns:
          type: array
          items: { $ref: '#/components/schemas/DataEntityColumn' }
        relatedColumns:
          type: array
          items: { $ref: '#/components/schemas/DataEntityColumn' }
        relations:
          type: array
          items: { $ref: '#/components/schemas/DataRelationship' }
        relatedTables:
          type: array
          items: { type: string }
    DataEntityColumn:
      type: object
      description: Columna persistida en `system_data_field_catalog`, poblada desde information_schema.
      properties:
        columnId: { type: string }
        dataEntityId: { type: string, nullable: true }
        schemaName: { type: string }
        tableName: { type: string }
        columnName: { type: string }
        ordinalPosition: { type: integer, nullable: true }
        businessName: { type: string, nullable: true }
        dataType: { type: string, nullable: true }
        sqlDataType: { type: string, nullable: true }
        isNullable: { type: boolean }
        isPrimaryKey: { type: boolean }
        isForeignKey: { type: boolean }
        referencesEntityId: { type: string, nullable: true }
        referencesSchema: { type: string, nullable: true }
        referencesTable: { type: string, nullable: true }
        referencesColumnName: { type: string, nullable: true }
        defaultValue: { type: string, nullable: true }
        businessDescription: { type: string, nullable: true }
        technicalDescription: { type: string, nullable: true }
        systemPurpose: { type: string, nullable: true }
        businessPurpose: { type: string, nullable: true }
        containsPii: { type: boolean }
        piiType: { type: string, nullable: true }
        containsSensitive: { type: boolean }
        containsFinancial: { type: boolean }
        containsFinancialData: { type: boolean }
        containsRiskData: { type: boolean }
        usedInScoring: { type: boolean }
        usedInMl: { type: boolean }
        isMlCandidate: { type: boolean }
        validationRule: { type: object, nullable: true }
        allowedValues: { type: array, nullable: true, items: {} }
        status: { type: string, enum: [ACTIVE, DEPRECATED_CANDIDATE, DEPRECATED, DISABLED] }
        detectedFrom: { type: string }
        confidenceLevel: { type: string }
        reviewStatus: { type: string, enum: [AUTO_DETECTED, NEEDS_REVIEW, APPROVED, REJECTED] }
    DataRelationship:
      type: object
      description: Relacion entre tablas, normalmente derivada de FKs reales.
      properties:
        relationId: { type: string }
        relationshipId: { type: string }
        sourceDataEntityId: { type: string, nullable: true }
        targetDataEntityId: { type: string, nullable: true }
        sourceSchema: { type: string }
        sourceTable: { type: string }
        sourceColumn: { type: string, nullable: true }
        targetSchema: { type: string }
        targetTable: { type: string }
        targetColumn: { type: string, nullable: true }
        column: { type: string, nullable: true }
        referencesEntityId: { type: string, nullable: true }
        referencesColumn: { type: string, nullable: true }
        relationshipType: { type: string }
        cardinality: { type: string }
        optionality: { type: string }
        businessReason: { type: string, nullable: true }
        technicalReason: { type: string, nullable: true }
        confidenceLevel: { type: string }
        reviewStatus: { type: string, enum: [AUTO_DETECTED, NEEDS_REVIEW, APPROVED, REJECTED] }
    ToolRequirement:
      type: object
      description: Forma producida por `mapToolRequirement`.
      properties:
        requirementId: { type: string }
        endpointId: { type: string }
        toolId: { type: string }
        usageType: { type: string }
        isRequired: { type: boolean }
        failureImpact: { type: string }
        fallbackStrategy: { type: string }
        requiresMock: { type: boolean }
        requiresStressTest: { type: boolean }
        notes: { type: string, nullable: true }
        detectedFrom: { type: string }
        confidenceLevel: { type: string }
        reviewStatus: { type: string, enum: [AUTO_DETECTED, NEEDS_REVIEW, APPROVED, REJECTED] }
    DataImpact:
      type: object
      description: Forma producida por `mapDataImpact`.
      properties:
        impactId: { type: string }
        endpointId: { type: string }
        dataEntityId: { type: string }
        operationType: { type: string, example: 'READ | CREATE | UPDATE | DELETE' }
        impactLevel: { type: string }
        isPrimaryEntity: { type: boolean }
        isTransactional: { type: boolean }
        rollbackRequired: { type: boolean }
        affectsCustomerState: { type: boolean }
        affectsFinancialState: { type: boolean }
        affectsRiskState: { type: boolean }
        affectsLegalState: { type: boolean }
        affectsDeviceState: { type: boolean }
        affectsNotificationState: { type: boolean }
        requiresAuditLog: { type: boolean }
        requiresRegressionTest: { type: boolean }
        requiresStressTest: { type: boolean }
        notes: { type: string, nullable: true }
        detectedFrom: { type: string }
        confidenceLevel: { type: string }
        reviewStatus: { type: string, enum: [AUTO_DETECTED, NEEDS_REVIEW, APPROVED, REJECTED] }
    FieldImpact:
      type: object
      description: Forma producida por `mapFieldImpact`.
      properties:
        fieldImpactId: { type: string }
        endpointId: { type: string }
        dataEntityId: { type: string }
        fieldName: { type: string }
        fieldOperation: { type: string }
        isRequiredInput: { type: boolean }
        isGenerated: { type: boolean }
        isSensitive: { type: boolean }
        isMlCandidate: { type: boolean }
        mlFeatureGroup: { type: string, nullable: true }
        validationRule: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        confidenceLevel: { type: string }
        reviewStatus: { type: string, enum: [AUTO_DETECTED, NEEDS_REVIEW, APPROVED, REJECTED] }
...
    ReviewDecisionRequest:
      type: object
      description: Compartido por los endpoints `*/review` de `SystemsReviewController` (`reviewDecisionSchema`).
      properties:
        reviewStatus:
          type: string
          enum: [NEEDS_REVIEW, APPROVED, REJECTED]
        confidenceLevel:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        notes:
          type: string
          maxLength: 1000
          description: 'Solo se persiste en data_impact, field_impact y tool_requirement (NO en endpoints ni data_entities).'
      required: [reviewStatus]
...
    ReviewQueueResult:
      type: object
      properties:
        endpoints:
          type: object
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/Endpoint' } }
            total: { type: integer }
        dataEntities:
          type: object
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/DataEntity' } }
            total: { type: integer }
        dataEntityImpacts:
          type: object
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/DataImpact' } }
            total: { type: integer }
        fieldImpacts:
          type: object
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/FieldImpact' } }
            total: { type: integer }
        dataColumnImpacts:
          type: object
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/DataEntityColumn' } }
            total: { type: integer }
        toolRequirements:
          type: object
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/ToolRequirement' } }
            total: { type: integer }
...
    Envelope_EndpointDetail_EndpointOnly:
      type: object
      description: 'El endpoint de revisión (reviewEndpoint) devuelve el objeto Endpoint plano en data, no el envelope EndpointDetailResult.'
      properties:
        requestId: { type: string, nullable: true }
        timestamp: { type: string, format: date-time }
        data: { $ref: '#/components/schemas/Endpoint' }
...
    Envelope_DataEntity:
      type: object
      properties:
        requestId: { type: string, nullable: true }
        timestamp: { type: string, format: date-time }
        data: { $ref: '#/components/schemas/DataEntity' }
...
    Envelope_ReviewQueue:
      type: object
      properties:
        requestId: { type: string, nullable: true }
        timestamp: { type: string, format: date-time }
        data: { $ref: '#/components/schemas/ReviewQueueResult' }
    Envelope_ToolRequirement:
      type: object
      properties:
        requestId: { type: string, nullable: true }
        timestamp: { type: string, format: date-time }
        data: { $ref: '#/components/schemas/ToolRequirement' }
    Envelope_DataImpact:
      type: object
      properties:
        requestId: { type: string, nullable: true }
        timestamp: { type: string, format: date-time }
        data: { $ref: '#/components/schemas/DataImpact' }
    Envelope_FieldImpact:
      type: object
      properties:
        requestId: { type: string, nullable: true }
        timestamp: { type: string, format: date-time }
        data: { $ref: '#/components/schemas/FieldImpact' }
    Envelope_DataEntityColumn:
      type: object
      properties:
        requestId: { type: string, nullable: true }
        timestamp: { type: string, format: date-time }
        data: { $ref: '#/components/schemas/DataEntityColumn' }
...